Fix diacritics
The big churn is from bumping us to Harfbuzz 6.0.0 (we were on a master from June 2022, 6.0.0 was released December 2022) as it fixes some diacritic positioning with Arabic text.

The main fix, however, comes from marshaling the glyph offsets to Dart (and storing them in C++ which we weren't doing). This is a translation offset within the local space (baseline and advance) of the glyph. It was causing our Arabic diacritics (the "accents" if you will) to be positioned incorrectly.

Prior to the fix:
![image](https://user-images.githubusercontent.com/454182/235834041-a683851a-96b0-4c72-adb9-f68c97682c14.png)

Post fix:
![image](https://user-images.githubusercontent.com/454182/235834074-23db8fab-08be-4529-98b3-f8a04ebe623a.png)

Same word displayed in Google Fonts (confirming correctness, also confirmed by an Arabic reader  on our Discord):
![CleanShot 2023-05-02 at 21 43 33@2x](https://user-images.githubusercontent.com/454182/235834273-9e423c7f-9dc2-48c3-afec-c7f7a9986416.png)

Diffs=
a8c1e3e75 Fix diacritics (#5200)

+++ include/rive/text/text.hpp
@@ -37,7 +37,7 @@ class GlyphItr
 {
 public:
     GlyphItr() = default;
-    GlyphItr(const OrderedLine* line, const GlyphRun* run, size_t glyphIndex) :
+    GlyphItr(const OrderedLine* line, const GlyphRun* run, uint32_t glyphIndex) :
         m_line(line), m_run(run), m_glyphIndex(glyphIndex)
     {}
 
@@ -52,12 +52,12 @@ public:
 
     GlyphItr& operator++();
 
-    std::tuple<const GlyphRun*, size_t> operator*() const { return {m_run, m_glyphIndex}; }
+    std::tuple<const GlyphRun*, uint32_t> operator*() const { return {m_run, m_glyphIndex}; }
 
 private:
     const OrderedLine* m_line;
     const GlyphRun* m_run;
-    size_t m_glyphIndex;
+    uint32_t m_glyphIndex;
 };
 
 // Represents a line of text with runs ordered visually. Also tracks logical
@@ -91,23 +91,25 @@ private:
 
 public:
     const GlyphRun* lastRun() const { return m_runs.back(); }
-    size_t startGlyphIndex(const GlyphRun* run) const
+    uint32_t startGlyphIndex(const GlyphRun* run) const
     {
         switch (run->dir)
         {
             case TextDirection::ltr:
                 return m_startLogical == run ? m_line->startGlyphIndex : 0;
             case TextDirection::rtl:
-                return (m_endLogical == run ? m_line->endGlyphIndex : run->glyphs.size()) - 1;
+                return (m_endLogical == run ? m_line->endGlyphIndex
+                                            : (uint32_t)run->glyphs.size()) -
+                       1;
         }
         RIVE_UNREACHABLE();
     }
-    size_t endGlyphIndex(const GlyphRun* run) const
+    uint32_t endGlyphIndex(const GlyphRun* run) const
     {
         switch (run->dir)
         {
             case TextDirection::ltr:
-                return m_endLogical == run ? m_line->endGlyphIndex : run->glyphs.size();
+                return m_endLogical == run ? m_line->endGlyphIndex : (uint32_t)run->glyphs.size();
             case TextDirection::rtl:
                 return (m_startLogical == run ? m_line->startGlyphIndex : 0) - 1;
         }
